+++ Source/VelocityToCV.cpp
@@ -77,7 +77,7 @@ void VelocityToCV::PlayNote(NoteMessage note)
 
 double VelocityToCV::Value(int samplesIn)
 {
-   return ofMap(mVelocity, 0, 127, GetMin(), GetMax(), K(clamped));
+   return ofMap(static_cast<double>(mVelocity), 0, 127, GetMin(), GetMax(), K(clamped));
 }
 
 void VelocityToCV::SaveLayout(ofxJSONElement& moduleInfo)
